Ajout erreur sur delete album / artist

+++ src/main/java/com/java/eval/web/service/AlbumService.java
@@ -34,6 +34,9 @@ public class AlbumService {
 
     // Exercice 8
     public void deleteAlbum(Integer id) {
+        if ( !albumRepository.existsById(id) ) {
+            throw new EntityNotFoundException("L'album avec l'id : " + id + " n'existe pas !");
+        }
         albumRepository.deleteById(id);
     }
 }
